Include functions to set origin and background momentum in all grids.
dft_driver_setup_origin and dft_driver_setup_momentum may be used to set the
x0,y0,z0 and kx0,ky0,kz0 of all the grids.

Note that if one wants a relative momentum between grids (for example if
the impurity has a wavefunction and the liquid another) one of the two
must be set with cgrid3d_set_momentum

driver_(x/y/z)0 set to NAN initally so that driver_setup sets it to
default value nx/2 only if not assigned yet.